Use raw value represented types for enums that frequently change over time.
This includes (for now, at least):
- Expansion
- RaidSlug
- RegionSlug

The idea is to avoid the requirement for package updates for new expansions,
raids or similar. Users of the package should be able to define their own
constants or use other values they deem appropriate.

At the same time, this packages tries to help the users by providing known
constants.